requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pillow==10.4.0
numpy==1.26.4
torch==2.4.1
torchvision==0.19.1
scikit-learn==1.5.1
joblib==1.4.2